created an excpetion to throw if NumberLimit is less than 1
I needed to make a way to alert the user if they typed a number less than 1 for the range to guess from. I also created a test to verify the exception is thrown when the numberLimit is less than 1. Also changed the guessLimit to default to 1, to make it more challenging when guessing a number between 1 and 9.